test: add e2e tests for `https` option
