+++ test/helpers/run-browser.js
@@ -20,6 +20,7 @@ function runBrowser(config) {
     puppeteer
       .launch({
         headless: true,
+        ignoreHTTPSErrors: true,
         // args come from: https://github.com/alixaxel/chrome-aws-lambda/blob/master/source/index.js
         args: puppeteerArgs,
       })
